requirements.txt
streamlit==1.24.0
qrcode==7.4.2
pillow==10.0.0
pandas==2.0.3
streamlit-analytics==0.3.0
python-dotenv==1.0.0